logic finally worked

+++ CalculatorLogic/Calculator.cs
@@ -2,7 +2,7 @@
 
 public static class Calculator
 {   //Adding method of 2 numbers
-    public static int Add(int left, int right)
+    public static double Add(double left, double right)
     {
         return left + right;
     }
